bpf/btf: Move tracing BTF APIs to the BTF library
Move the BTF APIs used in tracing to the BTF library code for sharing it
with others.
Previously, to avoid complex dependency in a series I made it on the
tracing tree, but now it is a good time to move it to BPF tree because
these functions are pure BTF functions.

Signed-off-by: Masami Hiramatsu (Google) <mhiramat@kernel.org>